requirements.txt
numpy==1.26.4
streamlit==1.29.0
Pillow==10.2.0
keras==2.10.0
tensorflow==2.10.0